requirementt.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
dj-rest-auth==6.0.0
Django==4.2.15
django-rest-auth==0.9.5
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
msgpack==1.0.8
numpy==2.0.2
pandas==2.2.2
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
twisted==24.7.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.3
